requirements.txt
pandas
python-pptx
openpyxl
docx  
celery
flask
click
pyamqp